Speed up tests by tweaking sleeps
